feat: adiciona melhores validacoes ao usuario

+++ src/api/controllers/UserController.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express";
+import jwt from "jsonwebtoken";
 import { ValidationError } from "../errors/ValidationError";
 import { User } from "../../database/entities/User";
 import { UserService } from "../services/UserService";
@@ -49,7 +50,40 @@ export class UserController {
     const id = parseInt(req.params.id);
 
     try {
-      const user = await this.userService.getUserById(id);
+      const token = req.headers.authorization?.split(" ")[1];
+
+      if (!token) {
+        return res.status(401).json({
+          code: 401,
+          status: "Unauthorized",
+          message: "Token is required.",
+        });
+      }
+
+      const decodedToken = (token: string) => {
+        try {
+          const decoded = jwt.verify(
+            token,
+            "njnckmlazlnxidih83934g5j90vniejincb89233hjn2ivcieonihyvtzftg9xsinmc",
+          ) as { id: number };
+          return decoded;
+        } catch (error) {
+          return null;
+        }
+      };
+
+      const decoded = decodedToken(token);
+      if (!decoded) {
+        return res.status(400).json({
+          code: 400,
+          status: "Bad Request",
+          message: "Invalid Token.",
+        });
+      }
+
+      const authenticatedUserId = decoded.id;
+
+      const user = await this.userService.getUserById(id, authenticatedUserId);
 
       return res.status(200).json(user);
     } catch (error: unknown) {
@@ -79,8 +113,49 @@ export class UserController {
 
   async updateUser(req: Request, res: Response) {
     try {
-      const { name, cpf, birth, cep, email, password } = req.body;
+      const token = req.headers.authorization?.split(" ")[1];
+
+      if (!token) {
+        return res.status(401).json({
+          code: 401,
+          status: "Unauthorized",
+          message: "Token is required.",
+        });
+      }
+
+      const decodedToken = (token: string) => {
+        try {
+          const decoded = jwt.verify(
+            token,
+            "njnckmlazlnxidih83934g5j90vniejincb89233hjn2ivcieonihyvtzftg9xsinmc",
+          ) as { id: number };
+          return decoded;
+        } catch (error) {
+          return null;
+        }
+      };
+
+      const decoded = decodedToken(token);
+      if (!decoded) {
+        return res.status(400).json({
+          code: 400,
+          status: "Bad Request",
+          message: "Invalid Token.",
+        });
+      }
+
+      const authenticatedUserId = decoded.id;
       const id = parseInt(req.params.id);
+
+      if (id !== authenticatedUserId) {
+        return res.status(403).json({
+          code: 403,
+          status: "Forbidden",
+          message: "You are not authorized to update this user's information.",
+        });
+      }
+
+      const { name, cpf, birth, cep, email, password } = req.body;
 
       const UserData: Partial<User> = {
         name,
@@ -120,7 +195,48 @@ export class UserController {
 
   async deleteUser(req: Request, res: Response) {
     try {
+      const token = req.headers.authorization?.split(" ")[1];
+
+      if (!token) {
+        return res.status(401).json({
+          code: 401,
+          status: "Unauthorized",
+          message: "Token is required.",
+        });
+      }
+
+      const decodedToken = (token: string) => {
+        try {
+          const decoded = jwt.verify(
+            token,
+            "njnckmlazlnxidih83934g5j90vniejincb89233hjn2ivcieonihyvtzftg9xsinmc",
+          ) as { id: number };
+          return decoded;
+        } catch (error) {
+          return null;
+        }
+      };
+
+      const decoded = decodedToken(token);
+      if (!decoded) {
+        return res.status(400).json({
+          code: 400,
+          status: "Bad Request",
+          message: "Invalid Token.",
+        });
+      }
+
+      const authenticatedUserId = decoded.id;
       const id = parseInt(req.params.id);
+
+      if (id !== authenticatedUserId) {
+        return res.status(403).json({
+          code: 403,
+          status: "Forbidden",
+          message: "You are not authorized to delete this user.",
+        });
+      }
+
       await this.userService.deleteUser(id);
 
       return res.status(204).json();
